requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
altair==5.5.0
asn1crypto==1.5.1
attrs==25.3.0
backtesting==0.6.4
beautifulsoup4==4.13.4
blinker==1.9.0
bokeh==3.7.3
boto3==1.38.34
botocore==1.38.34
cachetools==5.5.2
ccxt==4.4.88
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.0.0
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
entrypoints==0.4
Faker==37.3.0
favicon==0.7.0
filelock==3.18.0
fonttools==4.58.2
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
htbuilder==0.9.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jwt==1.3.1
kiwisolver==1.4.8
lxml==5.4.0
Markdown==3.8
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.3
multidict==6.4.4
narwhals==1.41.1
numpy==2.3.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
prometheus_client==0.22.1
propcache==0.3.2
protobuf==5.29.5
pyarrow==18.1.0
pycares==4.8.0
pycparser==2.22
pydeck==0.9.1
PyJWT==2.10.1
pymdown-extensions==10.15
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyupbit==0.2.34
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
s3transfer==0.13.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.15.0
snowflake-snowpark-python==1.32.0
sortedcontainers==2.4.0
soupsieve==2.7
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.45.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.1
streamlit-faker==0.0.3
streamlit-image-coordinates==0.3.1
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
validators==0.35.0
websockets==15.0.1
wheel==0.45.1
xyzservices==2025.4.0
yarl==1.20.1
